resolved various warnings

+++ src/actors/npc/PigBoss.cpp
@@ -284,7 +284,8 @@ namespace pnk
                 else
                 {
                     // equal or lower
-                    v.y = -5.5 + (static_cast<double>(0.25f * _walkSpeed));
+                    v.y = -5.5f + 0.25f * _walkSpeed;
+//                    v.y = -5.5f + (static_cast<double>(0.25f * _walkSpeed));
                 }
 
                 if ((wp->_pos.x - hrc.x) * (wp->_pos.x - hrc.x) > 1600)  // long horizontal distance
